src/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.30
pydantic>=2.13.3
pydantic-settings>=2.14.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
google-generativeai>=0.8.6
grpcio>=1.80.0
email-validator==2.1.0
icalendar==6.1.0
pytest>=8.3.0
pytest-cov>=5.0.0
ruff>=0.6.0
